Naming, Organization, Documentation
