--- packages/@balancer-labs/contracts/solidity-utils/math/FixedPoint.sol
@@ -1,161 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-pragma solidity ^0.8.0;
-
-import "../../interfaces/solidity-utils/helpers/BalancerErrors.sol";
-
-import "./LogExpMath.sol";
-
-/* solhint-disable private-vars-leading-underscore */
-
-library FixedPoint {
-  uint256 internal constant ONE = 1e18; // 18 decimal places
-  uint256 internal constant TWO = 2 * ONE;
-  uint256 internal constant FOUR = 4 * ONE;
-  uint256 internal constant MAX_POW_RELATIVE_ERROR = 10000; // 10^(-14)
-
-  // Minimum base for the power function when the exponent is 'free' (larger than ONE).
-  uint256 internal constant MIN_POW_BASE_FREE_EXPONENT = 0.7e18;
-
-  function add(uint256 a, uint256 b) internal pure returns (uint256) {
-    // Fixed Point addition is the same as regular checked addition
-
-    uint256 c = a + b;
-    _require(c >= a, Errors.ADD_OVERFLOW);
-    return c;
-  }
-
-  function sub(uint256 a, uint256 b) internal pure returns (uint256) {
-    // Fixed Point addition is the same as regular checked addition
-
-    _require(b <= a, Errors.SUB_OVERFLOW);
-    uint256 c = a - b;
-    return c;
-  }
-
-  function mulDown(uint256 a, uint256 b) internal pure returns (uint256) {
-    uint256 product = a * b;
-    _require(a == 0 || product / a == b, Errors.MUL_OVERFLOW);
-
-    return product / ONE;
-  }
-
-  function mulUp(uint256 a, uint256 b) internal pure returns (uint256) {
-    uint256 product = a * b;
-    _require(a == 0 || product / a == b, Errors.MUL_OVERFLOW);
-
-    if (product == 0) {
-      return 0;
-    } else {
-      // The traditional divUp formula is:
-      // divUp(x, y) := (x + y - 1) / y
-      // To avoid intermediate overflow in the addition, we distribute the division and get:
-      // divUp(x, y) := (x - 1) / y + 1
-      // Note that this requires x != 0, which we already tested for.
-
-      return ((product - 1) / ONE) + 1;
-    }
-  }
-
-  function divDown(uint256 a, uint256 b) internal pure returns (uint256) {
-    _require(b != 0, Errors.ZERO_DIVISION);
-
-    if (a == 0) {
-      return 0;
-    } else {
-      uint256 aInflated = a * ONE;
-      _require(aInflated / a == ONE, Errors.DIV_INTERNAL); // mul overflow
-
-      return aInflated / b;
-    }
-  }
-
-  function divUp(uint256 a, uint256 b) internal pure returns (uint256) {
-    _require(b != 0, Errors.ZERO_DIVISION);
-
-    if (a == 0) {
-      return 0;
-    } else {
-      uint256 aInflated = a * ONE;
-      _require(aInflated / a == ONE, Errors.DIV_INTERNAL); // mul overflow
-
-      // The traditional divUp formula is:
-      // divUp(x, y) := (x + y - 1) / y
-      // To avoid intermediate overflow in the addition, we distribute the division and get:
-      // divUp(x, y) := (x - 1) / y + 1
-      // Note that this requires x != 0, which we already tested for.
-
-      return ((aInflated - 1) / b) + 1;
-    }
-  }
-
-  /**
-   * @dev Returns x^y, assuming both are fixed point numbers, rounding down. The result is guaranteed to not be above
-   * the true value (that is, the error function expected - actual is always positive).
-   */
-  function powDown(uint256 x, uint256 y) internal pure returns (uint256) {
-    // Optimize for when y equals 1.0, 2.0 or 4.0, as those are very simple to implement and occur often in 50/50
-    // and 80/20 Weighted Pools
-    if (y == ONE) {
-      return x;
-    } else if (y == TWO) {
-      return mulDown(x, x);
-    } else if (y == FOUR) {
-      uint256 square = mulDown(x, x);
-      return mulDown(square, square);
-    } else {
-      uint256 raw = LogExpMath.pow(x, y);
-      uint256 maxError = add(mulUp(raw, MAX_POW_RELATIVE_ERROR), 1);
-
-      if (raw < maxError) {
-        return 0;
-      } else {
-        return sub(raw, maxError);
-      }
-    }
-  }
-
-  /**
-   * @dev Returns x^y, assuming both are fixed point numbers, rounding up. The result is guaranteed to not be below
-   * the true value (that is, the error function expected - actual is always negative).
-   */
-  function powUp(uint256 x, uint256 y) internal pure returns (uint256) {
-    // Optimize for when y equals 1.0, 2.0 or 4.0, as those are very simple to implement and occur often in 50/50
-    // and 80/20 Weighted Pools
-    if (y == ONE) {
-      return x;
-    } else if (y == TWO) {
-      return mulUp(x, x);
-    } else if (y == FOUR) {
-      uint256 square = mulUp(x, x);
-      return mulUp(square, square);
-    } else {
-      uint256 raw = LogExpMath.pow(x, y);
-      uint256 maxError = add(mulUp(raw, MAX_POW_RELATIVE_ERROR), 1);
-
-      return add(raw, maxError);
-    }
-  }
-
-  /**
-   * @dev Returns the complement of a value (1 - x), capped to 0 if x is larger than 1.
-   *
-   * Useful when computing the complement for values with some level of relative error, as it strips this error and
-   * prevents intermediate negative values.
-   */
-  function complement(uint256 x) internal pure returns (uint256) {
-    return (x < ONE) ? (ONE - x) : 0;
-  }
-}
