+++ packages/@cryptozoo/ethereum/contracts/ContinuousTokenOfferingPool/pool-utils/interfaces/IBasePool.sol
