+++ packages/@cryptozoo/ethereum/contracts/ContinuousTokenOfferingPool/pool-utils/interfaces/IMinimalSwapInfoPool.sol
