--- packages/@cryptozoo/ethereum/contracts/ERC20ManagedSupply/test/MockERC20ManagedSupply.sol
@@ -1,21 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-pragma solidity ^0.8.0;
-
-import "../ERC20ManagedSupply.sol";
-
-contract MockERC20ManagedSupply is ERC20ManagedSupply {
-  constructor(string memory _name, string memory _symbol) ERC20(_name, _symbol) {}
-}
