+++ packages/@cryptozoo/ethereum/contracts/solidity-utils/helpers/TemporarilyPausable.sol
@@ -1,0 +1,136 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+pragma solidity ^0.8.0;
+
+import "@balancer-labs/ethereum/contracts/interfaces/solidity-utils/helpers/BalancerErrors.sol";
+import "@balancer-labs/ethereum/contracts/interfaces/solidity-utils/helpers/ITemporarilyPausable.sol";
+
+/**
+ * @dev Allows for a contract to be paused during an initial period after deployment, disabling functionality. Can be
+ * used as an emergency switch in case a security vulnerability or threat is identified.
+ *
+ * The contract can only be paused during the Pause Window, a period that starts at deployment. It can also be
+ * unpaused and repaused any number of times during this period. This is intended to serve as a safety measure: it lets
+ * system managers react quickly to potentially dangerous situations, knowing that this action is reversible if careful
+ * analysis later determines there was a false alarm.
+ *
+ * If the contract is paused when the Pause Window finishes, it will remain in the paused state through an additional
+ * Buffer Period, after which it will be automatically unpaused forever. This is to ensure there is always enough time
+ * to react to an emergency, even if the threat is discovered shortly before the Pause Window expires.
+ *
+ * Note that since the contract can only be paused within the Pause Window, unpausing during the Buffer Period is
+ * irreversible.
+ */
+abstract contract TemporarilyPausable is ITemporarilyPausable {
+  // The Pause Window and Buffer Period are timestamp-based: they should not be relied upon for sub-minute accuracy.
+  // solhint-disable not-rely-on-time
+
+  uint256 private constant _MAX_PAUSE_WINDOW_DURATION = 90 days;
+  uint256 private constant _MAX_BUFFER_PERIOD_DURATION = 30 days;
+
+  uint256 private immutable _pauseWindowEndTime;
+  uint256 private immutable _bufferPeriodEndTime;
+
+  bool private _paused;
+
+  constructor(uint256 pauseWindowDuration, uint256 bufferPeriodDuration) {
+    _require(pauseWindowDuration <= _MAX_PAUSE_WINDOW_DURATION, Errors.MAX_PAUSE_WINDOW_DURATION);
+    _require(bufferPeriodDuration <= _MAX_BUFFER_PERIOD_DURATION, Errors.MAX_BUFFER_PERIOD_DURATION);
+
+    uint256 pauseWindowEndTime = block.timestamp + pauseWindowDuration;
+
+    _pauseWindowEndTime = pauseWindowEndTime;
+    _bufferPeriodEndTime = pauseWindowEndTime + bufferPeriodDuration;
+  }
+
+  /**
+   * @dev Reverts if the contract is paused.
+   */
+  modifier whenNotPaused() {
+    _ensureNotPaused();
+    _;
+  }
+
+  /**
+   * @dev Returns the current contract pause status, as well as the end times of the Pause Window and Buffer
+   * Period.
+   */
+  function getPausedState()
+    external
+    view
+    override
+    returns (
+      bool paused,
+      uint256 pauseWindowEndTime,
+      uint256 bufferPeriodEndTime
+    )
+  {
+    paused = !_isNotPaused();
+    pauseWindowEndTime = _getPauseWindowEndTime();
+    bufferPeriodEndTime = _getBufferPeriodEndTime();
+  }
+
+  /**
+   * @dev Sets the pause state to `paused`. The contract can only be paused until the end of the Pause Window, and
+   * unpaused until the end of the Buffer Period.
+   *
+   * Once the Buffer Period expires, this function reverts unconditionally.
+   */
+  function _setPaused(bool paused) internal {
+    if (paused) {
+      _require(block.timestamp < _getPauseWindowEndTime(), Errors.PAUSE_WINDOW_EXPIRED);
+    } else {
+      _require(block.timestamp < _getBufferPeriodEndTime(), Errors.BUFFER_PERIOD_EXPIRED);
+    }
+
+    _paused = paused;
+    emit PausedStateChanged(paused);
+  }
+
+  /**
+   * @dev Reverts if the contract is paused.
+   */
+  function _ensureNotPaused() internal view {
+    _require(_isNotPaused(), Errors.PAUSED);
+  }
+
+  /**
+   * @dev Reverts if the contract is not paused.
+   */
+  function _ensurePaused() internal view {
+    _require(!_isNotPaused(), Errors.NOT_PAUSED);
+  }
+
+  /**
+   * @dev Returns true if the contract is unpaused.
+   *
+   * Once the Buffer Period expires, the gas cost of calling this function is reduced dramatically, as storage is no
+   * longer accessed.
+   */
+  function _isNotPaused() internal view returns (bool) {
+    // After the Buffer Period, the (inexpensive) timestamp check short-circuits the storage access.
+    return block.timestamp > _getBufferPeriodEndTime() || !_paused;
+  }
+
+  // These getters lead to reduced bytecode size by inlining the immutable variables in a single place.
+
+  function _getPauseWindowEndTime() private view returns (uint256) {
+    return _pauseWindowEndTime;
+  }
+
+  function _getBufferPeriodEndTime() private view returns (uint256) {
+    return _bufferPeriodEndTime;
+  }
+}
