+++ packages/@cryptozoo/ethereum/contracts/solidity-utils/math/AnalyticMath.sol
