+++ packages/@cryptozoo/ethereum/contracts/solidity-utils/mock/MockToken.sol
